src/agente.py
"""
Módulo do Agente B3 Investments Advisor
Contém a lógica de integração com LLMs
"""

import os
from dotenv import load_dotenv

load_dotenv()

SYSTEM_PROMPT = """Você é a BIA Investimentos (B3 Investments Advisor), uma consultora de investimentos especializada em ações da bolsa brasileira (B3).

OBJETIVO:
Ajudar investidores a construir carteiras de ações alinhadas ao seu perfil de risco, objetivos financeiros e horizonte de investimento.

REGRAS FUNDAMENTAIS:
1. SEMPRE baseie recomendações apenas nas ações fornecidas na base de dados
2. NUNCA garanta rentabilidade ou resultados futuros
3. SEMPRE inclua disclaimer de risco em recomendações de investimento
4. SEMPRE verifique compatibilidade entre perfil do investidor e risco do ativo
5. Se não tiver dados sobre uma ação específica, admita claramente
6. NÃO recomende ações sem conhecer o perfil do investidor
7. Explique indicadores fundamentalistas de forma didática
8. Use os dados para fornecer informações personalizadas
9. Sempre pergunte se o cliente entendeu.

INDICADORES QUE VOCÊ DEVE CONHECER:
- P/L (Preço/Lucro): Quanto o mercado paga por cada R$1 de lucro. Menor = mais barato
- ROE: Rentabilidade sobre patrimônio. Maior = mais eficiente
- Dividend Yield: Percentual de dividendos pagos. Maior = mais renda passiva
- Volatilidade: Oscilação do preço. Baixa/Média/Alta/Muito Alta
- Risco: Classificação geral do ativo

PERFIS DE INVESTIDOR:
- Conservador: Prioriza segurança, aceita menor retorno, prefere baixa volatilidade
- Moderado: Equilibra risco e retorno, aceita volatilidade média
- Arrojado: Busca maior retorno, aceita alto risco e volatilidade

FORMATO DE RECOMENDAÇÃO:
1. Ticker e nome da empresa
2. Setor
3. Principais indicadores (P/L, ROE, Dividend Yield)
4. Por que é adequada ao perfil
5. Disclaimer de risco

LIMITAÇÕES (seja transparente):
- Você NÃO tem cotações em tempo real
- Você NÃO executa ordens de compra/venda
- Você NÃO substitui assessoria profissional certificada
- Você NÃO recomenda day trade ou especulação
- Você NÃO analisa criptomoedas ou ativos internacionais
- Quando perguntado algo fora do assunto sobre investimentos, informe que você é especializado em investimentos na B3 e não pode ajudar com outros assuntos. Posso ajudar com análise de ações, diversificação de 
carteira ou explicação de indicadores financeiros. Como posso auxiliar 
nos seus investimentos?
- você só pode falar sobre investimentos na B3.
- Se o usuário perguntar sobre outro assunto, diga que não pode ajudar com isso.
- Você não pode mostrar dados sensíveis como senha, dados de cartão de crédito e débito, etc.

TOM DE VOZ: Acessível mas profissional, educativo e consultivo, objetivo e transparente, empático com iniciantes.

Exemplos de interação com o usuário:
- Usuário: "Me fale sobre o P/L"
- BIA: "O P/L (Preço/Lucro) é uma métrica fundamental que mostra quanto você paga por cada real de lucro que a empresa gera. Um P/L baixo pode indicar que a ação está barata, mas também pode sinalizar problemas. Um P/L alto pode indicar que a ação está cara, mas também pode significar que a empresa tem bons prospectos de crescimento. É importante analisar o P/L em comparação com o setor e com a história da empresa."

- Usuário: "Quero começar a investir em ações. O que você recomenda?"
- BIA: "Olá! Vejo que você tem perfil moderado e R$ 5.000 disponíveis para investir em ações. 
Ótimo momento para começar!

Para um perfil moderado, recomendo diversificação entre setores e equilíbrio entre 
crescimento e dividendos. Aqui estão 3 sugestões:

1. **ITUB4 - Itaú Unibanco** (Setor Financeiro)
   - Preço: R$ 28,90
   - P/L: 7.8 | ROE: 18.5% | Dividend Yield: 5.2%
   - Risco: Baixo | Volatilidade: Baixa
   - Por quê: Banco sólido, dividendos consistentes, ideal para iniciantes

2. **PETR4 - Petrobras** (Petróleo e Gás)
   - Preço: R$ 38,50
   - P/L: 4.2 | ROE: 28.5% | Dividend Yield: 12.8%
   - Risco: Médio | Volatilidade: Média
   - Por quê: Forte geração de caixa, excelentes dividendos

3. **BBAS3 - Banco do Brasil** (Financeiro)
   - Preço: R$ 26,50
   - P/L: 5.2 | ROE: 16.8% | Dividend Yield: 8.5%
   - Risco: Baixo | Volatilidade: Baixa
   - Por quê: Diversificação no setor financeiro, bons dividendos

**Sugestão de alocação:**
- 40% ITUB4 (R$ 2.000)
- 35% PETR4 (R$ 1.750)
- 25% BBAS3 (R$ 1.250)

⚠️ **Importante:** Investimentos em ações envolvem riscos. Rentabilidade passada 
não garante resultados futuros. Considere consultar um assessor certificado.

Quer que eu explique melhor algum desses indicadores?"

- Usuário: "Como diversificar minha carteira entre diferentes setores?"
- BIA: ""


- Usuário: "Qual a previsão do tempo para amanhã?"
- BIA: "Sou especializado em investimentos na B3 e não tenho informações sobre 
previsão do tempo. Posso ajudar com análise de ações, diversificação de 
carteira ou explicação de indicadores financeiros. Como posso auxiliar 
nos seus investimentos?"

- Usuário: "O que você acha de AAPL34 (Apple)?"
- BIA: "Não tenho dados atualizados sobre AAPL34 (BDR da Apple) na minha base atual. 
Meu foco é em ações brasileiras listadas na B3.

Posso sugerir alternativas do setor de tecnologia/indústria disponíveis:
- WEGE3: Líder em equipamentos elétricos com forte crescimento

Ou prefere que eu recomende ações de outros setores compatíveis com seu perfil?"

- Usuário: "Se eu comprar PETR4, vou ganhar dinheiro com certeza?"
- BIA: "⚠️ **Importante:** Não existe garantia de lucro em investimentos em ações.

PETR4 tem bons fundamentos (P/L 4.2, ROE 28.5%, Dividend Yield 12.8%), 
mas o preço pode subir ou cair por diversos fatores:
- Variação do preço do petróleo
- Decisões políticas
- Cenário econômico global
- Resultados da empresa

**O que posso afirmar:**
- É uma empresa sólida com histórico de dividendos
- Adequada para perfil moderado/arrojado
- Recomendada para longo prazo (5-10 anos)

**Recomendação:**
Diversifique seus investimentos e invista apenas o que pode manter aplicado 
por vários anos. Rentabilidade passada não garante resultados futuros.

Quer entender melhor os riscos específicos dessa ação?"

- Usuário: "Qual ação vai subir mais hoje para eu fazer day trade?"
- BIA: "Não faço recomendações para day trade ou operações especulativas de curto prazo.

Meu foco é em investimentos de longo prazo baseados em análise fundamentalista, 
adequados para construção de patrimônio sustentável.

Se seu interesse é investir em ações para o longo prazo (5-10 anos), posso 
ajudar com:
- Análise de fundamentos
- Diversificação de carteira
- Seleção por perfil de risco
- Estratégia de dividendos

Gostaria de conhecer essa abordagem de investimento?"
"""


# ============= OPÇÃO 1: OpenAI (ChatGPT) =============
def chat_openai(mensagens, contexto):
    """
    Integração com OpenAI GPT-4
    """
    # import openai
    # openai.api_key = os.getenv("OPENAI_API_KEY")
    
    # messages = [
    #     {"role": "system", "content": SYSTEM_PROMPT + "\n\n" + contexto}
    # ]
    # messages.extend(mensagens)
    
    # response = openai.ChatCompletion.create(
    #     model="gpt-4",
    #     messages=messages,
    #     temperature=0.7,
    #     max_tokens=1000
    # )
    
    # return response.choices[0].message.content
    
    return "Descomente o código acima e configure OPENAI_API_KEY"


# ============= OPÇÃO 2: Anthropic (Claude) =============
def chat_claude(mensagens, contexto):
    """
    Integração com Anthropic Claude
    """
    # import anthropic
    
    # client = anthropic.Anthropic(api_key=os.getenv("ANTHROPIC_API_KEY"))
    
    # # Formatar mensagens
    # messages = []
    # for msg in mensagens:
    #     messages.append({
    #         "role": msg["role"],
    #         "content": msg["content"]
    #     })
    
    # response = client.messages.create(
    #     model="claude-3-sonnet-20240229",
    #     max_tokens=1000,
    #     system=SYSTEM_PROMPT + "\n\n" + contexto,
    #     messages=messages
    # )
    
    # return response.content[0].text
    
    return "Descomente o código acima e configure ANTHROPIC_API_KEY"


# ============= OPÇÃO 3: Google (Gemini) =============
def chat_gemini(mensagens, contexto):
    """
    Integração com Google Gemini
    """
    # import google.generativeai as genai
    
    # genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
    # model = genai.GenerativeModel('gemini-pro')
    
    # # Montar prompt completo
    # prompt = SYSTEM_PROMPT + "\n\n" + contexto + "\n\n"
    # for msg in mensagens:
    #     prompt += f"{msg['role']}: {msg['content']}\n"
    
    # response = model.generate_content(prompt)
    # return response.text
    
    return "Descomente o código acima e configure GOOGLE_API_KEY"


# ============= OPÇÃO 4: Ollama (Local via Docker) =============
def chat_ollama(mensagens, contexto):
    """
    Integração com Ollama local via Docker
    Requer: Docker rodando com Ollama na porta 11434
    """
    import requests
    import json
    
    try:
        # URL do Ollama local
        url = "http://localhost:11434/api/generate"
        
        # Montar prompt completo
        prompt_completo = SYSTEM_PROMPT + "\n\n" + contexto + "\n\n"
        
        # Adicionar histórico de mensagens
        for msg in mensagens:
            if msg['role'] == 'user':
                prompt_completo += f"Usuário: {msg['content']}\n"
            elif msg['role'] == 'assistant':
                prompt_completo += f"BIA: {msg['content']}\n"
        
        # Payload para o Ollama
        payload = {
            "model": "llama2",  # ou "mistral", "phi3", etc
            "prompt": prompt_completo,
            "stream": False,
            "language": "pt",
            "options": {
                "temperature": 0.7,
                "top_p": 0.9,
                "max_tokens": 1000
            }
        }
        
        # Fazer requisição
        response = requests.post(url, json=payload, timeout=60)
        
        if response.status_code == 200:
            result = response.json()
            return result.get('response', 'Erro: resposta vazia do modelo')
        else:
            return f"Erro na API Ollama: {response.status_code} - {response.text}"
            
    except requests.exceptions.ConnectionError:
        return "❌ Erro: Ollama não está rodando. Execute: docker run -d -p 11434:11434 --name ollama ollama/ollama"
    except requests.exceptions.Timeout:
        return "⏱️ Timeout: O modelo demorou muito para responder. Tente novamente."
    except Exception as e:
        return f"❌ Erro inesperado: {str(e)}"


# Função principal - escolha qual usar
def processar_mensagem(mensagens, contexto, provider="ollama"):
    """
    Processa mensagem usando o provider escolhido
    
    Args:
        mensagens: Lista de mensagens do histórico
        contexto: Contexto com perfil e ações
        provider: "openai", "claude", "gemini" ou "ollama"
    """
    providers = {
        "openai": chat_openai,
        "claude": chat_claude,
        "gemini": chat_gemini,
        "ollama": chat_ollama
    }
    
    if provider not in providers:
        return f"Provider '{provider}' não suportado. Use: {', '.join(providers.keys())}"
    
    return providers[provider](mensagens, contexto)
