src/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Descomente a API que você for usar:
# openai>=1.0.0
# anthropic>=0.7.0
# google-generativeai>=0.3.0
